requirements.txt
streamlit
dashscope
numpy
faiss-cpu